Allow promotions from a loss that makes SPRT pass certain
The best network doesn't have to be the network that loses a match for the challenger to promote because challenger losing on the 400th match can satisfy both SPRT pass and the 55% win rate. Fix up promotion check to ignore who wins or loses and just make sure the current best is being challenged, and if the promotion criteria is met, promote the challenger.